requirements.txt
aiohttp
azure-functions
azure-storage-blob
fastapi[all]
python-multipart